pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "face-rec-heatmaps"
version = "0.1.0"
description = "Minimal face recognition with landmark heatmaps and margin heads."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
dependencies = [
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "timm>=0.9.12",
  "mediapipe>=0.10.14",
  "dlib>=19.24.4",
  "numpy>=1.24.0",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "tqdm>=4.66.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.3.0",
  "tensorboard>=2.14.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"
